requirements/runtime.txt
matplotlib
mmcv>=1.1.1
torchvision
numpy
opencv-contrib-python
Pillow
terminaltables
tensorboard
tqdm
